requirements.txt
PyOpenGL==3.1.10
glfw==2.10.0
numpy==2.4.0
pygame